refactor: tag component
