+++ src/components/boxes/tag/TagComponent.tsx
@@ -1,23 +1,33 @@
+"use client"
+
 import React, { FC } from 'react';
 import styles from './TagComponent.module.css';
 
 interface TagComponentProps {
-  text: string;
   id: number;
-  onRemove: (id: number) => void;
+  text: string;
+  setTags: React.Dispatch<React.SetStateAction<TagType[]>>;
 }
 
-const TagComponent: FC<TagComponentProps> = ({ text, id, onRemove }) => {
-  const handleRemove = () => {
-    onRemove(id);
+const TagComponent: FC<TagComponentProps> = ({ id, text, setTags }) => {
+
+  const handleTagRemove = () => {
+    setTags(prevTags => prevTags.filter(tag => tag.id !== id));
   };
+
 
   return (
     <div className={styles.tag}>
       {text}
-      <button type='button' className={styles.removeButton} onClick={handleRemove} aria-label="Remove Tag">
+      <button
+        type="button"
+        className={styles.removeButton}
+        onClick={handleTagRemove}
+        aria-label="Remove Tag"
+      >
         &times;
       </button>
+
     </div>
   );
 };
